+++ src/main/java/com/nhnacademy/bookpub/bookpubfront/review/dto/response/GetProductReviewInfoResponseDto.java
@@ -3,7 +3,7 @@ package com.nhnacademy.bookpub.bookpubfront.review.dto.response;
 import lombok.Getter;
 
 /**
- * Some description here.
+ * 상품의 상품평 요약정보를 조회할 때 사용되는 Dto.
  *
  * @author : 정유진
  * @since : 1.0
@@ -11,5 +11,5 @@ import lombok.Getter;
 @Getter
 public class GetProductReviewInfoResponseDto {
     private Long reviewCount;
-    private Long productStar;
+    private Integer productStar;
 }
